+++ csa-app/model/application.go
@@ -41,7 +41,7 @@ type Application struct {
 	PrimaryLanguage   string            `gorm:"-" json:"-" yaml:"-"`
 	PrimaryLangCount  int               `gorm:"-" json:"-" yaml:"-"`
 	InfoFindings      int               `json:"infoFindings"`
-	RawScore          int               `json:"rawScore"`
+	RawScore          int64             `json:"rawScore"`
 	RawContainerScore int               `json:"rawContainerScore"`
 	RawCloudScore     int               `json:"rawCloudScore"`
 	NumCrits          int               `json:"numCrits"`
